obix-chatbot-backend/requirements-docker.txt
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
gunicorn==21.2.0
PyJWT==2.10.0
PyYAML==6.0.2
cryptography==43.0.3
bcrypt==4.2.1
openai==0.27.10
requests==2.32.3
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.8.2
pydantic==2.10.1
pydantic-settings==2.6.1
langchain==0.3.8
langchain-core==0.3.21
langchain-community==0.3.8
dj-database-url==2.1.0
whitenoise==6.6.0
google-generativeai==0.8.3
mistralai==0.4.2 